+++ CHANGELOG.rst
@@ -8,8 +8,10 @@ Features:
 
 - Add ``fields.Pluck`` for serializing a single field from a nested object (:issue:`800`). Thanks :user:`timc13` for the
   feedback and :user:`deckar01` for the implementation.
-- **Backwards-incompatible**: Passing a string argument as ``only`` to ``fields.Nested`` is no longer supported.
+- *Backwards-incompatible*: Passing a string argument as ``only`` to ``fields.Nested`` is no longer supported.
   Use ``fields.Pluck`` instead (:issue:`800`).
+- Raise a `StringNotCollectionError` if ``only`` or ``exclude`` is
+  passed as a string to ``fields.Nested`` (:pr:`931`).
 
 3.0.0b13 (2018-08-04)
 +++++++++++++++++++++