+++ CHANGELOG.rst
@@ -17,6 +17,8 @@ Bug fixes:
 
 - Raise ``ValidationError`` instead of ``TypeError`` when non-iterable types are
   validated with ``many=True`` (:issue:`851`). Thanks :user:`tuukkamustonen`.
+- Return ``[]`` as ``ValidationError.valid_data`` instead of ``{}`` with
+  ``many=True`` (:issue:`907`). Thanks :user:`tuukkamustonen`.
 
 3.0.0b13 (2018-08-04)
 +++++++++++++++++++++